refactor: notion-cache → notion-context, env vars all-optional, OAuth-first
리뷰 4건 (PR #1, .opencode/INSTALL.md L8/L25/L26/L28) 한 번에 반영:

- INSTALL.md plugin git URL 의 owner 를 minjun0219 로 박음
- AGENT_TOOLKIT_NOTION_MCP_URL 을 옵션화하고 기본값을 Notion 공식
  remote MCP (https://mcp.notion.com/mcp) 로 둠
- AGENT_TOOLKIT_NOTION_MCP_TOKEN 제거 — Notion remote MCP 가 OAuth
  로 인증하므로 plugin 단에서 bearer token 을 다루지 않음
- 캐시 기본 디렉터리를 ~/.cache/notion-context/pages 로 옮김

부수 작업:
- lib/notion-cache.ts → lib/notion-context.ts (history 보존 git mv)
- README, AGENTS, plugin / cache 코드 / 테스트 import 경로 동기화

검증: bun run typecheck clean, bun run test 16/16 pass.
