+++ README.md
@@ -14,8 +14,8 @@ opencode 전용 plugin. Notion 페이지를 캐시 우선으로 읽는 도구 3 
 │       ├── agent-toolkit.ts        # plugin entrypoint + 도구 3 개
 │       └── agent-toolkit.test.ts
 ├── lib/
-│   ├── notion-cache.ts             # TTL 파일 캐시 + key 정규화 + normalize
-│   └── notion-cache.test.ts
+│   ├── notion-context.ts           # TTL 파일 캐시 + key 정규화 + normalize
+│   └── notion-context.test.ts
 ├── skills/
 │   └── notion-spec-reader/SKILL.md # Notion → 한국어 스펙 정리 skill
 ├── .mcp.json                        # context7 MCP 등록 (개발 보조용)
@@ -31,20 +31,21 @@ opencode 전용 plugin. Notion 페이지를 캐시 우선으로 읽는 도구 3 
 `opencode.json` 의 `plugin` 배열에 추가하고 opencode 를 재시작:
 
 ```json
-{ "plugin": ["agent-toolkit@git+https://github.com/<owner>/coding-agent-toolkit.git"] }
+{ "plugin": ["agent-toolkit@git+https://github.com/minjun0219/coding-agent-toolkit.git"] }
 ```
 
 자세한 환경변수 / 검증 흐름은 [`.opencode/INSTALL.md`](./.opencode/INSTALL.md) 참고.
 
 ## 환경변수
 
-| 변수 | 필수 | 설명 |
+전부 옵션. 기본값을 바꿔야 할 때만 설정한다.
+
+| 변수 | 기본값 | 설명 |
 | --- | --- | --- |
-| `AGENT_TOOLKIT_NOTION_MCP_URL` | ✅ | remote Notion MCP base URL (`POST {url}/getPage` 가정) |
-| `AGENT_TOOLKIT_NOTION_MCP_TOKEN` | ⛔️ | bearer token |
-| `AGENT_TOOLKIT_NOTION_MCP_TIMEOUT_MS` | ⛔️ | 기본 `15000` |
-| `AGENT_TOOLKIT_CACHE_DIR` | ⛔️ | 기본 `.agent-cache/notion/pages` |
-| `AGENT_TOOLKIT_CACHE_TTL` | ⛔️ | 초 단위, 기본 `86400` |
+| `AGENT_TOOLKIT_NOTION_MCP_URL` | `https://mcp.notion.com/mcp` | remote Notion MCP base URL. 인증은 OAuth 가 처리하므로 토큰 변수는 없다. |
+| `AGENT_TOOLKIT_NOTION_MCP_TIMEOUT_MS` | `15000` | remote 호출 timeout (ms) |
+| `AGENT_TOOLKIT_CACHE_DIR` | `~/.cache/notion-context/pages` | 페이지 캐시 디렉터리 |
+| `AGENT_TOOLKIT_CACHE_TTL` | `86400` | 캐시 TTL (초) |
 
 ## 도구
